Fixes from feedback
